delete,update pages done

+++ src/Pages/MyListings.jsx
@@ -1,12 +1,13 @@
 import { getAuth } from 'firebase/auth';
 import React, { useEffect, useState } from 'react';
 import PostData from '../Components/PostData';
+import Swal from 'sweetalert2';
 
 
 const MyListings = () => {
     const [posts, setPosts] = useState([]);
     const user = getAuth().currentUser;
-    console.log(posts);
+    //console.log(posts);
 
     useEffect(() => {
         if (user) {
@@ -14,12 +15,40 @@ const MyListings = () => {
         }
     }, [])
 
-    const handleDelete = (id) =>{
-        console.log(id);
-        fetch(`http://localhost:3000/deletePost/${id}`,{
-            method: 'DELETE'
-        }).then(res => res.json()).then(data => console.log(data))
+    const handleDelete = (id) => {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to delete this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:3000/deletePost/${id}`, {
+                    method: 'DELETE',
+                    headers:{
+                        "Content-type" : "application/json"
+                    }
+                }).then(res => res.json()).then(data => {
+                    console.log(data);
+                    if (data.deletedCount) {
+                        const remainingData = posts.filter(post => post._id !== id);
+                        setPosts(remainingData);
+                        console.log(remainingData);
+                    }
+                })
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success"
+                });
+            }
+        });
+
     }
+    
 
     return (
         <div>
@@ -39,7 +68,7 @@ const MyListings = () => {
                     </thead>
                     <tbody>
                         {
-                            posts.map((post,index) => <PostData key={index} index={index} post={post} handleDelete={handleDelete}></PostData>)
+                            posts.map((post, index) => <PostData key={index} index={index} post={post} handleDelete={handleDelete}></PostData>)
                         }
                     </tbody>
                 </table>
